+++ tree.jquery.debug.js
@@ -3005,10 +3005,6 @@ var jqtree = (function (exports) {
         this._refreshElements(null);
         return this._element;
       }
-      refreshHitAreas() {
-        this._dndHandler.refresh();
-        return this._element;
-      }
       reload(onFinished) {
         this._doLoadDataFromUrl(null, null, onFinished);
         return this._element;
@@ -3137,7 +3133,6 @@ var jqtree = (function (exports) {
         const loadData = this.loadData.bind(this);
         const openNode = this._openNodeInternal.bind(this);
         const refreshElements = this._refreshElements.bind(this);
-        const refreshHitAreas = this.refreshHitAreas.bind(this);
         const selectNode = this.selectNode.bind(this);
         const treeElement = this._htmlElement;
         const triggerEvent = this._triggerEvent.bind(this);
@@ -3149,6 +3144,9 @@ var jqtree = (function (exports) {
         const isNodeSelected = selectNodeHandler.isNodeSelected.bind(selectNodeHandler);
         const removeFromSelection = selectNodeHandler.removeFromSelection.bind(selectNodeHandler);
         const getMouseDelay = () => this.options.startDndDelay ?? 0;
+        const refreshHitAreas = () => {
+          dndHandler.refresh();
+        };
         const dataLoader = new DataLoader({
           dataFilter,
           loadData,